third_party/blink/web_tests/platform/webview/external/wpt/css/css-anchor-position/idlharness-expected.txt

This is a testharness.js-based test.
[FAIL] CSSPositionFallbackRule interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface object length
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface object name
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] CSSPositionFallbackRule interface: attribute name
  assert_own_property: self does not have own property "CSSPositionFallbackRule" expected property "CSSPositionFallbackRule" missing
[FAIL] Stringification of positionFallback
  assert_class_string: class string of positionFallback expected "[object CSSPositionFallbackRule]" but got "[object Null]"
[FAIL] CSSPositionFallbackRule interface: positionFallback must inherit property "name" with the proper type
  assert_inherits: provided value is not an object
[FAIL] CSSTryRule interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface object length
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface object name
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule interface: attribute style
  assert_own_property: self does not have own property "CSSTryRule" expected property "CSSTryRule" missing
[FAIL] CSSTryRule must be primary interface of try1
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try1 is not defined"
[FAIL] Stringification of try1
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try1 is not defined"
[FAIL] CSSTryRule interface: try1 must inherit property "style" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try1 is not defined"
[FAIL] CSSTryRule must be primary interface of try2
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try2 is not defined"
[FAIL] Stringification of try2
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try2 is not defined"
[FAIL] CSSTryRule interface: try2 must inherit property "style" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: try2 is not defined"
Harness: the test ran to completion.
